+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Current (3.0.7)
+* FR1: YalseeLayout: common layout for all pages (#355)
+
 ### Version 3.0.6 (18/06/2021)
 Updates